PeaShooter Working

+++ src/sample/Bomb.java
@@ -4,7 +4,7 @@ public class Bomb extends Plant
 {
     private pair Range; // +- xr , +- yr
     private String Type ; // CherryBomb , PotatoMine
-    public Bomb(int H,int x, int y, String type, int xr, int yr,int c) { super(H,x,y,c); Type=type; Range = new pair(xr,yr);}
+    public Bomb(double H,double x, double y, String type, double xr, double yr,double c) { super(H,x,y,c); Type=type; Range = new pair(xr,yr);}
     public String getType() { return Type; }
     public void setType(String type) { Type = type; }
     public pair getRange() { return Range; }
