+++ src/sample/Character.java
@@ -2,10 +2,10 @@ package sample;
 
 public abstract class Character
 {
-    private int Health;
+    private double Health;
     private pair Position;
-    public Character(int H,int x, int y) { Health = H ; Position = new pair(x,y); }
-    public int getHealth() { return Health; }
-    public void setHealth(int health) { Health = health; }
+    public Character(double H,double x, double y) { Health = H ; Position = new pair(x,y); }
+    public double getHealth() { return Health; }
+    public void setHealth(double health) { Health = health; }
     public pair getPosition() { return Position; }
 }
