+++ src/sample/LawnMower.java
@@ -3,7 +3,7 @@ package sample;
 public class LawnMower extends Character
 {
     boolean Used;
-    public LawnMower(int x, int y) { super(100,x,y);Used = false; }
+    public LawnMower(double x, double y) { super(100,x,y);Used = false; }
     public void Use() {Used = true;}
     public boolean isUsed() { return Used; }
 }
