+++ src/sample/Plant.java
@@ -2,6 +2,7 @@ package sample;
 
 public class Plant extends Character
 {
-    private int Cost;
-    public Plant(int H,int x, int y,int c) { super(H, x, y); Cost = c;}
+    private double Cost;
+
+    public Plant(double H,double x, double y,double c) { super(H, x, y); Cost = c;}
 }
